fix filenames
